fix(root): rebase onboarding on App, add hook which control shownOnboarding
